Refactor: 分層架構
presentation-domain-data
