+++ app/src/main/AndroidManifest.xml
@@ -13,7 +13,7 @@
         android:theme="@style/Theme.StarterBookOrganizer"
         tools:targetApi="31">
         <activity
-            android:name=".AddBookActivity"
+            android:name=".presentation.add_book.AddBookActivity"
             android:exported="false"
             android:label="@string/title_activity_add_book"
             android:theme="@style/Theme.StarterBookOrganizer" />
