+++ app/src/main/java/com/example/starterbookorganizer/presentation/add_book/Category.kt
@@ -1,0 +1,10 @@
+package com.example.starterbookorganizer.presentation.add_book
+
+/**
+ * Created by AidenChang on 2025/9/1
+ */
+enum class Category {
+    READ,
+    READING,
+    WISHLIST
+}
